+++ libs/xui/src/layout/index.ts
@@ -1,1 +1,6 @@
-export * from './public-api';
+export * from './layout.module';
+export * from './layout.component';
+export * from './header.component';
+export * from './sider.component';
+export * from './content.component';
+export * from './footer.component';