+++ libs/xui/src/layout/public-api.ts
@@ -1,6 +1,0 @@
-export * from './layout.module';
-export * from './layout.component';
-export * from './header.component';
-export * from './sider.component';
-export * from './content.component';
-export * from './footer.component';