+++ libs/xui/src/progress/progress.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { XuiProgressComponent } from './progress.component';
-import { XuiIconModule } from '../icon';
+import { XuiDecagramModule } from '../decagram';
 
 @NgModule({
   declarations: [XuiProgressComponent],
-  imports: [CommonModule, XuiIconModule],
+  imports: [CommonModule, XuiDecagramModule],
   exports: [XuiProgressComponent]
 })
 export class XuiProgressModule {}