+++ libs/xui/src/select-old/select.types.ts
@@ -1,9 +1,0 @@
-export type XuiSelectModeType = 'default' | 'multiple' | 'tags';
-
-export interface XuiSelectOption {
-  label: string | number | null;
-  value: any;
-  disabled?: boolean;
-  hide?: boolean;
-  groupLabel?: string | number;
-}