+++ libs/xui/src/slider/slider.component.ts
@@ -1,72 +1,95 @@
 import {
+  AfterViewInit,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
+  ElementRef,
   Input,
+  OnChanges,
   OnInit,
   Optional,
   Self,
+  SimpleChanges,
+  ViewChild,
   ViewEncapsulation
 } from '@angular/core';
 import { ControlValueAccessor, NgControl } from '@angular/forms';
-import { TranslateService } from '@ngx-translate/core';
-import { InputBoolean } from '../utils';
+import { XuiTooltipComponent } from '../tooltip/tooltip.component';
+import { inNextTick, InputBoolean, InputNumber } from '../utils';
+import { BehaviorSubject, map } from 'rxjs';
 
 @Component({
   selector: 'xui-slider',
   exportAs: 'xuiSlider',
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush,
-  templateUrl: 'slider.component.html'
+  templateUrl: 'slider.component.html',
+  host: {
+    '(click)': 'move($event.x)'
+  }
 })
-export class XuiSliderComponent implements ControlValueAccessor, OnInit {
-  _value?: string;
+export class XuiSliderComponent implements ControlValueAccessor, OnInit, AfterViewInit, OnChanges {
+  _posX = new BehaviorSubject<number>(0);
+  _value?: number;
   touched = false;
-  onChange = (source?: string) => {};
+  onChange = (source?: number) => {};
   onTouched = () => {};
 
-  @Input() placeholder?: string;
-  @Input() @InputBoolean() disabled: boolean = false;
-  @Input() color: 'primary' | 'primary-alt' | 'secondary' | 'error' | 'warning' | 'success' | 'minimal' | string =
-    'primary';
-  @Input() secondColor?: 'primary' | 'primary-alt' | 'secondary' | 'error' | 'success' | 'minimal' | string;
-  @Input() size: 'normal' | 'small' = 'normal';
-  @Input() dataList?: string[] | null;
-  @Input() type: 'text' | 'password' | 'color' | 'date' | 'email' | 'number' = 'text';
+  position$ = this._posX.pipe(map(x => ({ x, y: 0 })));
+
+  @ViewChild('track', { static: true }) trackElm!: ElementRef;
+  @ViewChild('tooltipRef') tooltipRef!: XuiTooltipComponent;
+
+  // TODO: default and focus
+  // @Input() @InputBoolean() disabled: boolean = false;
+  @Input() @InputBoolean() tooltipDisabled: boolean = false;
+  @Input() color: 'primary' | 'primary-alt' | 'secondary' | 'error' | 'warning' | 'success' | string = 'primary';
+  @Input() secondColor?: 'primary' | 'primary-alt' | 'secondary' | 'error' | 'success' | string;
+  // @Input() size: 'normal' | 'small' = 'normal';
+
+  @Input() @InputNumber() min: number = 0;
+  @Input() @InputNumber() max: number = 100;
+  @Input() @InputNumber() step: number = 1;
+  @Input() tipFormatter?: (value: number) => string;
+  @Input() marks?: SliderMark[];
+  @Input() @InputBoolean() range: boolean = false;
 
   @Input()
   get value() {
-    return this._value;
+    return this._value ?? this.min;
   }
 
   set value(v) {
     if (this._value !== v) {
       this._value = v;
+
       this.onChange(v);
       this.changeDetectorRef.markForCheck();
     }
   }
 
-  get tooltip() {
-    return '42 %';
+  private get percentage() {
+    return this.getPercentage(this.value);
   }
 
-  // get styles() {
-  //   return `xui-input xui-input-${this.color} xui-input-${this.size}`;
-  // }
+  get width() {
+    return 100 - this.percentage;
+  }
+
+  get tooltip() {
+    return String(this.tipFormatter ? this.tipFormatter(this.value) : this.value);
+  }
 
   getColor(color: string) {
     return `xui-slider-color-${color}`;
   }
 
-  get errorMessage() {
-    const msg = this.control?.getError('message');
-    return this.translation.instant(msg);
+  getPercentage(absolute: number) {
+    return ((absolute - this.min) / (this.max - this.min)) * 100;
   }
 
   constructor(
     private changeDetectorRef: ChangeDetectorRef,
-    private translation: TranslateService,
     @Self() @Optional() public control?: NgControl
   ) {
     if (this.control) {
@@ -78,20 +101,18 @@ export class XuiSliderComponent implements ControlValueAccessor, OnInit {
     this.control?.statusChanges!.subscribe(() => this.changeDetectorRef.markForCheck());
   }
 
-  get invalid(): boolean {
-    return !!this.control?.invalid;
+  async ngAfterViewInit() {
+    await inNextTick();
+    this.setPositionByPercentage(this.percentage);
   }
 
-  get showError(): boolean {
-    if (!this.control) {
-      return false;
+  ngOnChanges(changes: SimpleChanges) {
+    if (changes['value']) {
+      this.setPositionByPercentage(this.percentage);
     }
-
-    const { dirty, touched } = this.control;
-    return this.invalid ? dirty! || touched! : false;
   }
 
-  writeValue(source: string) {
+  writeValue(source: number) {
     this.value = source;
   }
 
@@ -109,4 +130,35 @@ export class XuiSliderComponent implements ControlValueAccessor, OnInit {
       this.touched = true;
     }
   }
+
+  async move(screenX: number) {
+    const cursorOffset = 5;
+    const rect = this.hostRect;
+    const width = rect.width - 10;
+
+    const pos = Math.min(Math.max(screenX - rect.x - cursorOffset, 0), width);
+    const value = (pos / width) * (this.max - this.min); // 0 <-> max - min
+    const newValue = Math.round(value / this.step) * this.step;
+
+    this.value = newValue + this.min;
+    this.setPositionByPercentage(this.percentage);
+    await inNextTick();
+
+    this.tooltipRef.show(0);
+  }
+
+  private get hostRect() {
+    return this.trackElm.nativeElement.getBoundingClientRect() as DOMRect;
+  }
+
+  private setPositionByPercentage(value: number) {
+    this._posX.next((value / 100) * this.hostRect.width - 5);
+    this.changeDetectorRef.markForCheck();
+  }
+}
+
+export interface SliderMark {
+  label: string;
+  value: number;
+  color?: string; // TODO
 }